requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-community==0.0.20
pypdf==3.17.0
python-dotenv==1.0.0
python-docx==1.1.0
streamlit-option-menu==0.3.2
scikit-learn==1.3.0
numpy==1.24.0